Update the code to avoid errors
I noticed that when the machine generated a 0 the first time, the code stopped because the first condition was true. I implemented a solution based on a Do While type of loop so the first time the code always goes through even if the machine gives a 0 as a first answer.